requirements.txt
python >= 3.7
pytest
requests
requests-mock
pyyaml